Allow forwarding arguments to `wasm-pack`
The motivation here is to pass the `--dev` flag to newer versions of `wasm-pack` since that will do nice things like skip size optimizations. This introduces a dependency on `clap` for more robust cli arg handling (including `--help` output), rather than the previous strategy of reading from the args directly.

While here I noticed that if it failed to listen to the port, the `cargo watch` subcommand would continue to run, so I had it kill that process in that case.